Move FetchURL() and FetchURLToFile() to binaryfetcher package
